backend/requirements.txt
Flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
tensorflow==2.16.1
opencv-python==4.10.0.84
numpy==1.26.4
Flask-SocketIO==5.3.6
python-socketio==5.11.2
PyJWT==2.8.0
